seperate axios function calls to seperate file called cats.ts
