+++ src/api/query-client.ts
@@ -1,0 +1,9 @@
+// api/query-client.ts
+
+import { QueryClient } from "@tanstack/react-query";
+
+const defaultQueryConfig = { staleTime: 60000 };
+
+export const queryClient = new QueryClient({
+  defaultOptions: { queries: defaultQueryConfig },
+});
