+++ src/api/useCat.ts
@@ -3,11 +3,7 @@ import { axios } from './axios';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
 import { getQueryKey } from './queryKey';
 import { CatTributes } from './types';
-
-export async function getCats() {
-  const { data } = await axios.get('/images/search');
-  return data;
-}
+import { getCats } from './cats'
 
 export function useCats(page: number) {
   const query = useQuery<CatTributes[], Error>(
