+++ src/api/useVote.ts
@@ -1,25 +1,8 @@
 import { axios } from './axios';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import assert from 'assert';
 import { getQueryKey } from './queryKey';
 import { VoteProps } from './types';
-
-export async function postCatVotes(catVote: VoteProps) {
-  assert(
-    process.env.REACT_APP_VOTE_URL,
-    'env variable not set:process.env.REACT_APP_VOTE_URL'
-  );
-  assert(
-    process.env.REACT_APP_API_KEY,
-    'env variable not set:process.env.REACT_APP_API_KEY'
-  );
-
-  return await axios
-    .post('/votes', catVote
-    )
-    .then((res) => console.log(res))
-    .catch((err) => console.error(err));
-}
+import { postCatVotes } from './cats'
 
 export function useVote(catVote: VoteProps) {
   const queryClient = useQueryClient();
